Remove split crates
