+++ Cargo.toml
@@ -1,7 +1,15 @@
-[workspace]
-members = [
-  "zeta",
-  "zeta-core",
-  "zeta-irc",
-  "zeta-plugins",
-]
+[package]
+name = "zeta"
+version = "0.1.0"
+authors = ["Mikkel Kroman <mk@maero.dk>"]
+edition = "2018"
+
+[dependencies]
+argh = "0.1.10"
+miette = { version = "5.5.0", features = ["fancy"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_yaml = "0.9"
+thiserror = "1.0"
+tokio = { version = "1.0", features = ["full"] }
+tracing = "0.1"
+tracing-subscriber = "0.3"
