--- config.yaml
@@ -1,23 +1,0 @@
-production: &base
-  networks:
-    - &uplink
-      url: irc://znc.zeta:6667
-      nickname: zeta
-      password: mk:password
-      channels:
-        - '#uplink'
-
-development:
-  <<: *base
-
-  networks:
-    - &test
-      url: irc://localhost:6667
-      nickname: zeta-test
-      channels:
-        - '#test'
-    - &test2
-      url: irc://localhost:6667
-      nickname: zeta-test2
-      channels:
-        - '#test'
