--- k8s/deployment.yml
@@ -1,31 +1,0 @@
-# Create the Zeta deployment and container
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: zeta
-  namespace: zeta
-  labels:
-    app: zeta
-    environment: production
-spec:
-  selector:
-    matchLabels:
-      app: zeta
-      environment: production
-  # Template for the created pods
-  template:
-    metadata:
-      namespace: zeta
-      labels:
-        app: zeta
-        environment: production
-    spec:
-      containers:
-      - name: zeta
-        env:
-          - name: RUST_LOG
-            value: zeta_core=debug
-        env:
-          - name: ZETA_ENV
-            value: production
-        image: mkroman/zeta:{{ build_commit }}
