--- sonar-project.properties
@@ -1,2 +1,0 @@
-sonar.projectKey=mkroman_zeta
-sonar.organization=mkroman
