--- zeta-core/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "zeta-core"
-version = "0.1.0"
-authors = ["Mikkel Kroman <mk@maero.dk>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bytes = "1"
-rand = "0.8.3"
-serde = { version = "1.0", features = ["derive"] }
-slab = "0.4.7"
-thiserror = "1.0"
-tokio = { version = "1.0", features = ["full"] }
-tokio-native-tls = "0.3.0"
-tokio-stream = { version = "0.1" }
-tracing = "0.1"
-url = { version = "2.1", features = ["serde"] }
