--- zeta-irc/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "zeta-irc"
-version = "0.1.0"
-authors = ["Mikkel Kroman <mk@maero.dk>"]
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-[features]
-default = ["builder"]
-
-# Enables the builders for the different types
-builder = []
-
-[dev-dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-rstest = "0.18"
