--- zeta-irc/src/frame.rs
@@ -1,2 +1,0 @@
-pub struct Frame {
-}
