--- zeta-plugins/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "zeta-plugins"
-version = "0.1.0"
-authors = ["Mikkel Kroman <mk@maero.dk>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lazy_static = "1.4"
