--- zeta-plugins/src/google_search.rs
@@ -1,9 +1,0 @@
-use crate::Plugin;
-
-pub struct GoogleSearch;
-
-impl Plugin for GoogleSearch {
-    fn new() -> GoogleSearch {
-        GoogleSearch {}
-    }
-}
