--- zeta-plugins/src/plugin.rs
@@ -1,5 +1,0 @@
-pub trait Plugin: Send + Sync {
-    fn new() -> Self
-    where
-        Self: Sized;
-}
