--- zeta/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "zeta"
-version = "0.1.0"
-authors = ["Mikkel Kroman <mk@maero.dk>"]
-edition = "2018"
-
-[dependencies]
-argh = "0.1.10"
-miette = { version = "5.5.0", features = ["fancy"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_yaml = "0.9"
-thiserror = "1.0"
-tokio = { version = "1.0", features = ["full"] }
-tracing = "0.1"
-tracing-subscriber = "0.3"
-zeta-core = { path = "../zeta-core", version = "*" }
